// client/src/Models/SignsOfPainsModel/SignsOfPainModel/Signs-Of-Pain-Model.ts
interface SignsOfPain {
  sex?: string;
  difficultySwallowingLiquids?: boolean;
  difficultrySwallowingSolidFood?: boolean;
  difficultySwallowingPills?: boolean;
  coughingWhileEating?: boolean;
  losingWeight?: boolean;
  ChokingDuringMeals?: boolean;
  AvoidingEatingWithOthers?: boolean;
  PainWhileSwallowing?: boolean;
  EnjoyingMeals?: boolean;
  feelingAnxiousAboutSwallowing?: boolean;
  numberOfSignsOfPain?: number;
  numberOfGoodFeeling?: number;
  language?: string;
}

export default SignsOfPain;